Update and rename Google_Earth_Engine_Script to Google_Earth_Engine_GPP_Script
This script is intended to export both halves of the GPP model: a tiff file containing all of the herbaceous data and a second file containing all of the woody data.  This script is inoperable without the 4 individual masks which will be available in the future as cml.tamu.edu.